+++ ceammc/ext/src/fx/fx_stutter.cpp
@@ -4,9 +4,9 @@
 constexpr auto PROP_MAXSIZE_MIN = 1;
 constexpr auto PROP_MAXSIZE_DEF = 1000;
 constexpr auto PROP_MAXSIZE_MAX = 10000;
-constexpr auto PROP_SIZE_MIN = PROP_MAXSIZE_MIN;
-constexpr auto PROP_SIZE_DEF = 50;
-constexpr auto PROP_SIZE_MAX = PROP_MAXSIZE_MAX;
+constexpr auto PROP_T_MIN = PROP_MAXSIZE_MIN;
+constexpr auto PROP_T_DEF = 50;
+constexpr auto PROP_T_MAX = PROP_MAXSIZE_MAX;
 
 namespace {
 size_t ms2samp(t_sample ms, t_sample sr)
@@ -20,7 +20,7 @@ FxStutter::FxStutter(const PdArgs& args)
     : SoundExternal(args)
     , state_(ST_PASS)
     , max_size_(nullptr)
-    , size_(nullptr)
+    , t_(nullptr)
     , clock_([this]() {
         state_ = ST_FROM_FX;
     })
@@ -35,23 +35,24 @@ FxStutter::FxStutter(const PdArgs& args)
     max_size_->setInitOnly();
     addProperty(max_size_);
 
-    size_ = new FloatProperty("@size", PROP_SIZE_DEF);
-    size_->checkClosedRange(PROP_SIZE_MIN, PROP_SIZE_MAX);
-    size_->setArgIndex(0);
-    addProperty(size_);
+    t_ = new FloatProperty("@t", PROP_T_DEF);
+    t_->checkClosedRange(PROP_T_MIN, PROP_T_MAX);
+    t_->setArgIndex(0);
+    t_->setUnits(PropValueUnits::MSEC);
+    addProperty(t_);
 }
 
 void FxStutter::initDone()
 {
     adjustBufferSize();
-    size_->checkClosedRange(PROP_SIZE_MIN, max_size_->value());
+    t_->checkClosedRange(PROP_T_MIN, max_size_->value());
 }
 
 void FxStutter::onInlet(size_t n, const AtomListView& lv)
 {
     if (lv.empty()) {
         state_ = ST_TO_FX;
-        clock_.delay(size_->value() * 2);
+        clock_.delay(t_->value() * 2);
     } else {
         clock_.unset();
         const bool on = lv.boolAt(0, false);
@@ -88,7 +89,7 @@ void FxStutter::setupDSP(t_signal** sig)
 void FxStutter::processBlock(const t_sample** in, t_sample** out)
 {
     const auto bs = blockSize();
-    const size_t n = std::min<size_t>(buffer_.size(), ms2samp(size_->value(), samplerate()));
+    const size_t n = std::min<size_t>(buffer_.size(), ms2samp(t_->value(), samplerate()));
 
     switch (state_) {
     case ST_PASS: